+++ app/src/main/java/woowacourse/payments/ui/registercard/ExpiryDateVisualTransformation.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.cardregister
+package woowacourse.payments.ui.registercard
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.OffsetMapping
