+++ app/src/test/java/woowacourse/payments/registercard/CardNumberVisualTransformationTest.kt
@@ -1,11 +1,11 @@
-package woowacourse.payments.cardregister
+package woowacourse.payments.registercard
 
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.input.VisualTransformation
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.payments.ui.cardregister.CardNumberVisualTransformation
+import woowacourse.payments.ui.registercard.CardNumberVisualTransformation
 
 class CardNumberVisualTransformationTest {
     private lateinit var visualTransformation: VisualTransformation
